Added ShoppingCartViewComponent to _Layout so that the shopping cart count can be populated when the user is logged in, and clears when it logs out. It uses Http.Context.Session
