+++ BulkyBookWeb/Areas/Customer/Controllers/HomeController.cs
@@ -23,13 +23,6 @@ namespace BulkyBookWeb.Areas.Customer.Controllers
 
         public IActionResult Index()
         {
-            var claimsIdentity = (ClaimsIdentity)User.Identity;
-            var claim = claimsIdentity.FindFirst(ClaimTypes.NameIdentifier);
-            if (claim != null)
-            {
-                HttpContext.Session.SetInt32(SD.SessionCart, _unitOfWork.ShoppingCart.GetAll(u => u.ApplicationUserId == claim.Value).Count());
-            }
-
             IEnumerable<Product> productList = _unitOfWork.Product.GetAll(includeProperties: "Category");
 
             return View(productList);
