+++ BulkyBookWeb/Views/Shared/Components/ShoppingCart/Default.cshtml
@@ -1,0 +1,3 @@
+﻿@model int 
+
+<i class="bi bi-cart"></i>&nbsp;(@Model)
